refactor: storage to use node filesystem operations
