+++ src/renderer/App.tsx
@@ -4,7 +4,7 @@ import { ErrorMessage } from './components/ErrorMessage';
 import { ItemList } from './components/ItemList';
 import { NewTodoItem } from './components/NewTodoItem';
 import { loadItems, saveItems, setupAppFolder } from './storage';
-import { Todo } from './types';
+import { Todo } from '../shared/types';
 import { addTodoItem, removeCompletedItems, updateTodoList } from './utils';
 
 const App: React.FC = () => {
