+++ src/renderer/components/ItemList/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Todo, TodoCallback } from '../../types';
+import type { Todo, TodoCallback } from '../../../shared/types';
 import { TodoItem } from '../TodoItem';
 
 type Props = {
